Refactor code, ChangeAvatarModal component created
